+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+- Update all dependencies to their latest versions
+- Update plugin code to use the latest share_plus API (SharePlus.instance.share with ShareParams)
+- Minor: Example app still has some deprecated withOpacity usage (does not affect plugin)
+
 ## 1.0.0
 
 - Initial release of flutter_long_screenshot plugin